Removed functions from server.js and put them in their own controllers. Updated .gitignore to ignore dbs/configs/.idea files
