+++ .gitignore
@@ -6,3 +6,6 @@ package-lock.json
 database.sqlite
 config.js
 .vscode
+.idea
+database.sqlite
+config.js
