+++ models/index.js
@@ -2,11 +2,7 @@ const sqlite3 = require('sqlite3')
 const {open} = require('sqlite')
 const config = require('./../config');
 
-
-
-
-module.exports = async (logger) => {
-    let db;
+async function validateDatabase(db, logger) {
     try {
         db = await open({
             filename: config.databaseFile,
@@ -22,12 +18,36 @@ module.exports = async (logger) => {
         logger.error(exception);
         process.exit();
     }
+}
+
+
+
+
+module.exports = async (logger) => {
+    let db;
+
+    try {
+        db = await open({
+            filename: config.databaseFile,
+            driver: sqlite3.Database,
+            mode: sqlite3.OPEN_READWRITE
+        });
+
+    } catch (exception) {
+        logger.error(exception);
+        process.exit();
+    }
+
+    // I'm not sure we _really_ need to wait for this, since it'll exit the application if it's invalid anyway.
+
+    await validateDatabase(db, logger);
 
 
     return {
         db,
         models: {
-            drivesModel: require('./drives')(db)
+            drivesModel: require('./drives')(db),
+            users: require('./users')(db),
         }
     }
 }
